requirements.txt
pyrogram==2.0.106
tgcrypto
uvloop
python-dotenv
aiohttp
Pillow
gTTS